% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_cc_insert_lastvalue.R
\name{cc_insert_lastvalue}
\alias{cc_insert_lastvalue}
\title{Insert the contents of `.Last.value` as a comment}
\usage{
cc_insert_lastvalue()
}
\value{
Character vector.
}
\description{
'.Last.value' is a built-in variable from the 'base' package that contains
the output of the last top-level expression that was run. This add-in function
grabs that output, turns it into a comment, and inserts it into the current
source document at the current cursor position. It also adds a datestamp.
}
